requirements.txt
autopep8
Django
pylint
djangorestframework
markdown
django-filter